+++ bridge/config/config.go
@@ -284,6 +284,7 @@ type BridgeValues struct {
 	Zulip              map[string]Protocol
 	Keybase            map[string]Protocol
 	Mumble             map[string]Protocol
+	Fluxer             map[string]Protocol
 	General            Protocol
 	Tengo              Tengo
 	Gateway            []Gateway