+++ gateway/bridgemap/bfluxer.go
@@ -1,0 +1,11 @@
+//go:build !nofluxer
+
+package bridgemap
+
+import (
+	bfluxer "github.com/matterbridge-org/matterbridge/bridge/fluxer"
+)
+
+func init() {
+	FullMap["fluxer"] = bfluxer.New
+}